# Conflicts:
#	zigbee-app-new/src/main/webapp/app/js/buygoodslist.js
#	zigbee-app-new/src/main/webapp/app/js/lookRdc.js
#	zigbee-app-new/src/main/webapp/app/js/rentRdc.js
#	zigbee-app-new/src/main/webapp/app/js/salegoodslist.js

+++ zigbee-app-new/src/main/webapp/app/js/buygoodslist.js
@@ -104,6 +104,14 @@ $().ready(function() {
         if(rdc.rentdate==undefined || rdc.rentdate==null || rdc.rentdate==0){
             usefulDate = daysRound+'天'
         }
+        var collectWords='<a class="fr noCollect" onclick="collect(this,'+rdc.id+')"><i class="iconfont">&#xe605;</i><em>收藏</em></a>';
+        if(rdc.collectUserIds && window.user){
+            for(var i=0;i<rdc.collectUserIds.length;i++){
+                if(rdc.collectUserIds[i]==window.user.id){
+                    collectWords='<a class="fr hasCollect" onclick="collect(this,'+rdc.id+')"><i class="iconfont">&#xe60c;</i><em>已收藏</em></a>';
+                }
+            }
+        }
         var score = [
             '<li class="imgCell"><a href="storehousedetail.html?id=' + rdc.id + '" onclick="getSoll()"><span>求购货源</span><div>'+
             '<p class="ellipsis">'+rdc.title+'</p><p class="position omg orange"><i class="iconfont">&#xe673;</i>'+rdc.sqm+'吨</p><span class="grab green">['+showTime+']</span>'+
@@ -111,17 +119,27 @@ $().ready(function() {
             '<p>有效期</p></div><div class="item"><h4>'+rdc.validEndTime+'</h4><p>报价截止日</p>'+
             '</div><div class="item"><h4 class="omg">'+rdc.username+'</h4><p>发布者</p></div></div></a>' +
 			'<div class="btnFn clearfix"><a href="storehousedetail.html?id='+rdc.id+'" class="fl"><i class="iconfont">&#xe65b;</i>查看</a>'+
-            '<a class="fr noCollect" onclick="collect(this)"><i class="iconfont">&#xe605;</i><em>收藏</em></a><a class="fr"><i class="iconfont">&#xe66c;</i>咨询</a></div></li>'
+            collectWords+'<a class="fr"><i class="iconfont">&#xe66c;</i>咨询</a></div></li>'
         ];
   		  return score.join("");
   	}
-    collect=function(ops) {
+    collect=function(ops,id) {
+        if(!(window.user && window.user.id!=0)){
+            layer.open({content: "请登入后收藏！", btn: '确定'});
+            return;
+        }
         var em = $(ops);
         if(em.hasClass('noCollect')){
+            $.post(ER.root+"/i/collect/addCollectRdc",{uid:window.user.id,collectId:id,collectType:2},function (data) {
+
+            });
             em.removeClass('noCollect').addClass('hasCollect');
             em.children('i').html('&#xe60c;');
             em.children('em').html('已收藏');
         }else{
+            $.post(ER.root+"/i/collect/delCollectById",{uid:window.user.id,collectId:id,collectType:2},function (data) {
+
+            });
             em.addClass('noCollect').removeClass('hasCollect');
             em.children('i').html('&#xe605;');
             em.children('em').html('收藏');
